Allow ALTER SYSTEM command to execute on whole cluster
A new unreserved keyword has been added to ALTER SYSTEM in order to
explicitly allow dispatching of the command to the segments. The keyword
is unsurprisingly named 'DISPATCH' and is not used elsewhere.

The main reason for being explicit for the distribution is that admins
should be able to execute alter system commands in a node level.

Needs documentation which is not included in this commit.