Minor vacuum command cleanups and refactor
	* Remove unused AO segment relations from vacuum_rel()
	* Special AO aux tables are non vacuumable, error out
	* Move the distinct phases of vacuum inside the tableam api
	* Perform all the distict phases of vacuum in single distributed transaction

The only part of the above that needs expanded explanation is the last.

Previously on master we were exectuting the three distinct phases in a mix of
local and distributed transactions. Please read discussion here:
https://groups.google.com/a/greenplum.org/forum/#!msg/gpdb-dev/lUwtNxaT3f0/Cg39CP8uAwAJ

This commit challenges the idea that the three phases will have to happen within
distinct distributed transactions. The pre and post phases are short enough in
comparisson to the compaction phase in order to cause any noticable delays in
the use of the vacuumed relations. Until this point, there does not seem to
exist any other valid justification for keeping the phases in seperate
transactions.

Co-authored-by: Asim R P <pasim@vmware.com>