Use StringInfo instead of a char in replace_string in pg_regress
StringInfo got available to frontend code also in commit 26aaf97b683 so it is
fine to use in pg_regress. This makes the code a little bit cleaner.

Although ecpg does not compile, try to keep it as up to date as possible. Due to
linking errors, stringinfo.h could not be included. Use a staticly defined
*_inline version instead.

However commit 26aaf97b683 is not yet included in 12.

Co-authored-by: Asim R P <pasim@vmware.com>